# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MathPlatform API",
        "description": "API for MathPlatform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/student/exams/{exam}": {
            "get": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Get exam questions",
                "description": "The exam can only be accessed after all active videos in the lesson are completed.",
                "operationId": "studentExamShow",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "description": "Exam ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam retrieved successfully"
                    },
                    "403": {
                        "description": "Exam is locked"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/student/exams/{exam}/start": {
            "post": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Start an exam",
                "description": "Starts an exam after all active lesson videos have been completed.",
                "operationId": "studentExamStart",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "description": "Exam ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam started successfully"
                    },
                    "403": {
                        "description": "Exam is locked"
                    },
                    "404": {
                        "description": "Exam not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/student/exams/{exam}/submit": {
            "post": {
                "tags": [
                    "Student Exams"
                ],
                "summary": "Submit exam answers",
                "description": "All questions must be answered. The backend calculates the score and determines whether the student passed.",
                "operationId": "studentExamSubmit",
                "parameters": [
                    {
                        "name": "exam",
                        "in": "path",
                        "description": "Exam ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attempt_id",
                                    "answers"
                                ],
                                "properties": {
                                    "attempt_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "question_id",
                                                "option_id"
                                            ],
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "option_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exam submitted successfully"
                    },
                    "403": {
                        "description": "Exam is locked"
                    },
                    "404": {
                        "description": "Exam not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Sanctum token returned by the admin login.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}